Make separate header file for recursive directory watch setup.
